requirements.txt
flask
pynacl